+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "java.project.sourcePaths": [
+    "Frontend/src"
+  ]
+}
