+++ Frontend/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "java.project.sourcePaths": ["src"],
+    "java.project.outputPath": "bin"
+}
