--- Frontend/Main_GUI.java
@@ -1,8 +1,0 @@
-import Main_Game.Main_Game_Page;
-import Main_Menu.Main_Menu_Page;
-
-public class Main_GUI {
-    public static void main(String[] args) {
-        new Main_Game_Page();
-    }
-}
