+++ Frontend/src/Main_GUI.java
@@ -1,0 +1,8 @@
+import main_game.Main_Game_Page;
+import main_menu.Main_Menu_Page;
+
+public class Main_GUI {
+    public static void main(String[] args) throws Exception {
+        new Main_Game_Page();
+    }
+}
