+++ Frontend/src/main_menu/Main_Menu_Page.java
@@ -1,4 +1,4 @@
-package Main_Menu;
+package main_menu;
 
 public class Main_Menu_Page {
